requirements.txt
GitPython
gradio